don't run in client server mode

+++ db4oj/test/src/com/db4o/test/acid/CrashSimulatingTest.java
@@ -30,6 +30,8 @@ public class CrashSimulatingTest {
     }
     
     public void test() throws IOException{
+    	
+    	if (Test.isClientServer()) return;
         
         new File(FILE).delete();
         new File(PATH).mkdirs();
